feat: staking list  (#80)
* remove old staking list and add new

* add new staking table with refresh and insert data

rename staking dialog to staking tab

* removing unnecessary code
improving naming
