+++ electrum/gui/qt/stake_dialog.py
@@ -43,32 +43,6 @@ from .create_new_stake_window import CreateNewStakingWindow
 from .staking_detail_tx_window import CompletedMultiClaimedStakeDialog
 from .terms_and_conditions_mixin import load_terms_and_conditions
 from .util import read_QIcon, WindowModalDialog, OkButton
-from ...stake import stake_api
-from .staking_list import staking_list
-
-
-def refresh_stake_dialog_window(window):
-    """
-    Call this function to refresh stake dialog window
-    TODO
-    """
-    current_staking_data = stake_api.get_detailed_stakes_data_for_addresses(
-        addresses=window.wallet.get_addresses()
-    )
-    current_height = window.wallet.get_local_height()
-
-    staking_list.insert_data(
-        table_data={
-            'Type': [get_verbal_type_name(stack_data=data) for data in current_staking_data],
-            'Start Date': [data['timestamp'] for data in current_staking_data],
-            'Amount': [data['staking_amount'] for data in current_staking_data],
-            'Staking Period': [data['staking_period'] for data in current_staking_data],
-            'Deposit Height': [data['deposit_height'] for data in current_staking_data],
-            'Blocks Left': [get_block_left(data, current_height) for data in current_staking_data],
-            'tx_hash': [data['tx_hash'] for data in current_staking_data],
-        },
-        context_menu_kwargs={'window': window, },
-    )
 
 
 def get_verbal_type_name(stack_data):
@@ -108,7 +82,7 @@ class CustomButton(QPushButton):
             self.func()
 
 
-def staking_dialog(window):
+def staking_tab(window):
     window.top_h_label = QHBoxLayout()
     window.create_stake_dialog = CreateNewStakingWindow(window)
 
@@ -133,8 +107,6 @@ def staking_dialog(window):
     window.top_h_label.addItem(verticalSpacer)
     window.top_h_label.addWidget(window.stake_balance_label)
 
-    window.staking_list = staking_list
-
     font = QFont()
     font.setUnderline(True)
     window.terms_button = QPushButton()
@@ -146,19 +118,18 @@ def staking_dialog(window):
     window.terms_button.setAutoDefault(True)
     window.terms_button.clicked.connect(terms_and_conditions_view)
 
-
-    w = QWidget()
-    vbox = QVBoxLayout(w)
+    widget = QWidget()
+    vbox = QVBoxLayout(widget)
 
     vbox.addStretch(1)
     vbox.addLayout(window.top_h_label)
+
     vbox.addWidget(window.staking_list)
+
     vbox.addWidget(window.terms_button)
     vbox.setStretchFactor(window.staking_list, 1000)
 
-    refresh_stake_dialog_window(window=window)
-
-    return w
+    return widget
 
 
 def terms_and_conditions_view():
